Refactor: Move repository and use case interfaces
Move repository and use case interfaces to separate subpackages
within the `domain.api` package for better organization.

- `domain.api.repository` for repository interfaces
- `domain.api.usecase` for use case interfaces

Delete `UsefulHabitEntryDto` as it's redundant with `EntryDto`.
Update `UsefulHabitsWithEntriesDto` to use `EntryDto`.

Rename `UsefulHabitWithEntries` entity to `UsefulHabitWithEntriesEntity`
for consistency.

Add `UsefulHabitWithEntriesRepository` and its implementation
`UsefulHabitWithEntriesRepositoryImpl`.
Add mappers for `UsefulHabitsWithEntriesDto` and
`UsefulHabitWithEntriesEntity`.

Update DI modules and relevant classes to reflect these changes.
Remove `UsefulHabitWithEntries` from `AppDatabase` entities as it's not a direct table.
