+++ app/src/main/java/com/kawunus/habityou/data/dto/UsefulHabitsWithEntriesDto.kt
@@ -2,5 +2,5 @@ package com.kawunus.habityou.data.dto
 
 class UsefulHabitsWithEntriesDto(
     val usefulHabit: UsefulHabitDto,
-    val entries: List<UsefulHabitEntryDto>
+    val entries: List<EntryDto>
 )
