+++ app/src/main/java/com/kawunus/habityou/data/impl/UsefulHabitRepositoryImpl.kt
@@ -2,7 +2,7 @@ package com.kawunus.habityou.data.impl
 
 import com.kawunus.habityou.data.database.dao.UsefulHabitDao
 import com.kawunus.habityou.data.dto.UsefulHabitDto
-import com.kawunus.habityou.domain.api.UsefulHabitsRepository
+import com.kawunus.habityou.domain.api.repository.UsefulHabitsRepository
 import com.kawunus.habityou.utils.mappers.toUsefulHabitDto
 import com.kawunus.habityou.utils.mappers.toUsefulHabitEntity
 import kotlinx.coroutines.flow.Flow
