+++ app/src/main/java/com/kawunus/habityou/di/RepositoryModule.kt
@@ -3,9 +3,11 @@ package com.kawunus.habityou.di
 import com.kawunus.habityou.data.impl.EntryRepositoryImpl
 import com.kawunus.habityou.data.impl.NoteRepositoryImpl
 import com.kawunus.habityou.data.impl.UsefulHabitRepositoryImpl
-import com.kawunus.habityou.domain.api.EntryRepository
-import com.kawunus.habityou.domain.api.NoteRepository
-import com.kawunus.habityou.domain.api.UsefulHabitsRepository
+import com.kawunus.habityou.data.impl.UsefulHabitWithEntriesRepositoryImpl
+import com.kawunus.habityou.domain.api.repository.EntryRepository
+import com.kawunus.habityou.domain.api.repository.NoteRepository
+import com.kawunus.habityou.domain.api.repository.UsefulHabitWithEntriesRepository
+import com.kawunus.habityou.domain.api.repository.UsefulHabitsRepository
 import org.koin.core.module.dsl.bind
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.module
@@ -17,4 +19,6 @@ val repositoryModule = module {
     singleOf(::UsefulHabitRepositoryImpl) { bind<UsefulHabitsRepository>() }
 
     singleOf(::EntryRepositoryImpl) { bind<EntryRepository>() }
+
+    singleOf(::UsefulHabitWithEntriesRepositoryImpl) { bind<UsefulHabitWithEntriesRepository>() }
 }
