+++ app/src/main/java/com/kawunus/habityou/domain/api/repository/UsefulHabitWithEntriesRepository.kt
@@ -1,0 +1,9 @@
+package com.kawunus.habityou.domain.api.repository
+
+import com.kawunus.habityou.data.dto.UsefulHabitsWithEntriesDto
+import kotlinx.coroutines.flow.Flow
+
+interface UsefulHabitWithEntriesRepository {
+
+    suspend fun getAllHabitsWithEntries(): Flow<List<UsefulHabitsWithEntriesDto>>
+}
