+++ app/src/main/java/com/kawunus/habityou/domain/usecase/DiaryInteractorImpl.kt
@@ -1,7 +1,7 @@
 package com.kawunus.habityou.domain.usecase
 
-import com.kawunus.habityou.domain.api.DiaryInteractor
-import com.kawunus.habityou.domain.api.NoteRepository
+import com.kawunus.habityou.domain.api.repository.NoteRepository
+import com.kawunus.habityou.domain.api.usecase.DiaryInteractor
 import com.kawunus.habityou.domain.model.Note
 import com.kawunus.habityou.utils.mappers.toNote
 import com.kawunus.habityou.utils.mappers.toNoteDto
