+++ app/src/main/java/com/kawunus/habityou/ui/diary/viewmodel/DiaryViewModel.kt
@@ -2,7 +2,7 @@ package com.kawunus.habityou.ui.diary.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.kawunus.habityou.domain.api.DiaryInteractor
+import com.kawunus.habityou.domain.api.usecase.DiaryInteractor
 import com.kawunus.habityou.domain.model.Note
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
